pyproject.toml
# https://betterprogramming.pub/a-pyproject-toml-developers-cheat-sheet-5782801fb3ed
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "network_exporter"
version = "1.35"
dependencies = [
  "setuptools==70.1.1",
  "fastapi==0.111.0",
  "pydantic==2.7.3",
  "uvicorn==0.30.1",
  "pan-python==0.17.0",
  "pan-os-python==1.12.1",
  "prometheus-client==0.20.0",
  "prometheus-fastapi-instrumentator==7.0.0",
  "xmltodict==0.13.0",
  "netmiko==4.3.0",
  "typer==0.12.3",
  "cachetools==5.5.0"
]
[tool.setuptools]
packages = ['src/network_exporter']

[project.optional-dependencies]
dev = [
  "pytest==8.2.2",
  "pytest-mock==3.14.0"
]
[project.scripts]
network-exporter = "network_exporter.runner:main"

[project.urls]
changelog = "./CHANGELOG.md"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
